+++ packages/jetbrains-plugin/src/test/resources/project-fixtures/basic-java-project-with-mongodb/src/main/java/alt/mongodb/javadriver/JavaDriverRepository.java
@@ -3,7 +3,12 @@ package alt.mongodb.javadriver;
 import com.mongodb.client.FindIterable;
 import com.mongodb.client.MongoClient;
 import com.mongodb.client.MongoCollection;
+import com.mongodb.client.model.Filters;
 import org.bson.Document;
+
+import java.util.ArrayList;
+import java.util.List;
+
 import static com.mongodb.client.model.Filters.*;
 
 public class JavaDriverRepository {
@@ -16,20 +21,10 @@ public class JavaDriverRepository {
         this.client = client;
     }
 
-    public FindIterable<Document> exampleFind() {
-        return getCollection().find(eq(AWARDS_WINS, 123));
-    }
-
-    public FindIterable<Document> exampleFindUsingCustomDSL() {
-        return findByField(524);
-    }
-
-    private MongoCollection<Document> getCollection() {
+    private List<Document> getGrade() {
         return client.getDatabase("sample_mflix")
-                .getCollection("movies");
-    }
-
-    private FindIterable<Document> findByField(int value) {
-        return getCollection().find(gt(IMDB_VOTES, value));
+                .getCollection("users")
+                .find(Filters.eq("email", 1))
+                .into(new ArrayList<>());
     }
 }