+++ packages/mongodb-dialects/java-driver/src/main/kotlin/com/mongodb/jbplugin/dialects/javadriver/glossary/JavaDriverDialect.kt
@@ -2,9 +2,13 @@ package com.mongodb.jbplugin.dialects.javadriver.glossary
 
 import com.intellij.psi.PsiElement
 import com.mongodb.jbplugin.dialects.Dialect
+import com.mongodb.jbplugin.dialects.DialectFormatter
 import com.mongodb.jbplugin.dialects.DialectParser
 
 object JavaDriverDialect : Dialect<PsiElement> {
     override val parser: DialectParser<PsiElement>
         get() = JavaDriverDialectParser
+
+    override val formatter: DialectFormatter
+        get() = JavaDriverDialectFormatter
 }