+++ packages/mongodb-dialects/java-driver/src/test/kotlin/com/mongodb/jbplugin/dialects/javadriver/glossary/JavaDriverDialectTest.kt
@@ -8,4 +8,9 @@ class JavaDriverDialectTest {
     fun `returns the correct parser`() {
         assertEquals(JavaDriverDialectParser, JavaDriverDialect.parser)
     }
+
+    @Test
+    fun `returns the correct formatter`() {
+        assertEquals(JavaDriverDialectFormatter, JavaDriverDialect.formatter)
+    }
 }