+++ packages/mongodb-mql-model/src/main/kotlin/com/mongodb/jbplugin/mql/CollectionSchema.kt
@@ -18,6 +18,29 @@ data class CollectionSchema(
         return recursivelyGetType(splitJsonPath, schema)
     }
 
+    fun allFieldNamesQualified(): List<Pair<String, BsonType>> = recursivelyAllFieldNamesQualified(emptyList(), schema)
+
+    private fun recursivelyAllFieldNamesQualified(
+        currentPath: List<String>,
+        root: BsonType,
+    ): List<Pair<String, BsonType>> =
+        when (root) {
+            is BsonArray -> recursivelyAllFieldNamesQualified(currentPath + "0", root.schema)
+            is BsonObject -> root.schema.flatMap { recursivelyAllFieldNamesQualified(currentPath + it.key, it.value) }
+            is BsonAnyOf ->
+                root.types
+                    .flatMap {
+                        recursivelyAllFieldNamesQualified(currentPath, it)
+                    }
+.fold(emptySet<BsonType>()) { acc, el ->
+                        acc + el.second
+                    }
+.let {
+                        listOf(Pair(currentPath.joinToString("."), BsonAnyOf(it)))
+                    }
+            else -> listOf(Pair(currentPath.joinToString("."), root))
+        }
+
     private fun recursivelyGetType(
         jsonPath: List<String>,
         root: BsonType,